chore: TypeScript 5.1 touchups
