+++ apps/automated/src/ui/repeater/repeaterItems-bindingToGestures-page.xml
@@ -1,8 +1,8 @@
 ﻿<Page xmlns="http://schemas.nativescript.org/tns.xsd" loaded="pageLoaded">
 	<Repeater id="repeater" items="{{ items }}">
 		<Repeater.itemTemplate>
-			<GridLayout width="94" height="94" >
-				<Label text="{{ $parents['Page'].parentViewProperty }}" tap="{{ tapItem }}" />
+			<GridLayout width="94" height="94">
+				<Label text="{{ text }}" tap="{{ tapItem }}" />
 			</GridLayout>
 		</Repeater.itemTemplate>
 	</Repeater>
