+++ packages/core/tsconfig.lib.json
@@ -4,7 +4,7 @@
 		"noEmitOnError": true,
 		"noEmitHelpers": true,
 		"declaration": true,
-		"noImplicitUseStrict": true,
+		"ignoreDeprecations": "5.0",
 		"removeComments": false,
 		"emitDecoratorMetadata": true,
 		"experimentalDecorators": true,
